refactor: set zeebe env vars in manifest
In order to have everything in one place and for better maintainability most of the env vars should be set in the statefulset manifest. This is done via this commit. The env vars can be calculated via kubernetes dependent environment variables, which are expanded at deploy time. The initial contact points can be calculated with helm template magic.