fix: move servicemonitor to root scope
Move servicemonitor template to root templates, and remove it from zeebe templates.
Move the values to root level, instead of having it in Zeebe section.

This should make clear that the servicemonitor is not only for zeebe but
also for other services and will scrape other services, like
zeebe-gateway, operate etc.

Make it possible to configure the servicemonitor and add documentation.