fix: use existingSecret string as secret value
If existingSecret contains a string, we expect that this is the secret value.
We use that to recreate the secret on upgrade. We reference our secret in the
identity deployment.

If the existingSecret contains anything else, like an object. We expect that it references an secret created by the user.

Expected format would be then in the values.yaml file:

existingSecret:
   name: secret-name-ref

Where secre-name-ref is pointing to a deployed secret in the same
namespace with that name.