feat: configure default mode for OpenShift
Configures the defaultMode of explicit volume mounts. This means
updating the Zeebe broker's `configMap.defaultMode` from `0744` (meaning
readable by all, writable and executable only by the owner) to `0755`
(meaning readable by all, writable and executable by the owner and the
group). This enables OpenShift deployments to execute shell scripts for
the broker (e.g. the `configMap`'s `startup.sh` script), as OpenShift
will run the container using an arbitrary UID but with the `root` or `0`
group.

Additionally configure the `defaultMode` for the curator cron job to be
`0744` (readable for all, everything else for the owner).