feat(openshift): add openshift profile
Adds getting-started guide for Openshift, including a workaround to
deploy the chart with Helm 3.2.x and higher, as well as some basic
technical documentation.

The documentation should be supplemented with an actual user guide
deployed on the official Camunda documentation. This README.md is meant
as a quick helper, not as actual complete documentation.