test(it): adds Openshift integration tests
Splits off the definition of the `integrationSuite` type into its own
file, leaving only the non-Openshift integration tests in
`integration_test.go`. Adds a new `openshift_test.go` which reuses the
`integrationSuite` type, but adds its own Openshift specific setup along
with only a basic end to end test.

The tests require the current context to be pointing to an Openshift
cluster, and for the user/service account to have the create
`projectrequests` permissions as well as the `get,list,watch` nodes.
The former makes it simpler to setup isolated projects (somewhat
equivalent to namespaces), and the latter is required by `terratest`
when checking if a service is available.