fix: align conditionals in identityPostgres
In the identity deployment, there is an environment variable named
IDENTITY_DATABASE_PASSWORD which is conditionally applied if
identityPostgresql.enabled is true, or if an externalDatabase is
configured. However, the referenced secret may not be created if
identityPostgresql is enabled, because the secret is only created if an
externalDatabase is configured.