fix: use correct application file (Operate)
Application file has been renamed on the way to a single application.
It is now called application.yaml, instead of application.yml.

If we still use the old name, we would create a spearate file and
not overwrite the existing one. This causes to load multiple configs,
spring will handle them in a way that the yaml overrides several configs
from the yml  file.

This causes that the application connected with the defaults, like to
localhost:9200, which of course doesn't work in the K8 environment.