fix: renovate should not update digests that default to empty string
A recent PR that introduced image digests causes errors in renovate
because renovate expects to have the digest be something other than an
empty string. And we do not want to mandate all customers have digests
by default because then the StatefulSet matchLabels may fail to match on
a helm-upgrade.