fix: resume alpha upgrades by renovate post-alpha6
in .github/renovate.json5, we have a hard-coded maximum alpha number
that we do not upgrade past. That number is alpha6, and we did not
expect to exceed this number because that represents 6 months after a
minor release (time for the next minor). But 8.8 is special in that we
started releasing 8.8 alpha's prior to the official minor release of
8.7.

I propose that we up this value to alpha9. Historically, we did not have
a maximum on the alpha number, but we found that application teams would
sometimes publish debug images under camunda/camunda:8.8-alpha99. We
wanted to avoid upgrades to such versions, which is also why adding a +
to the regex isn't the best solution here.