fix: remove the metadata in the namespace name
A recent change introduced extra metadata into the namespace name
https://github.com/camunda/camunda-platform-helm/pull/3675

One change that was made is that in the namespace name, the scenario is
added to the namespace.

This part of the change is not okay because it further increases the
likelihood of running into the 63 char limit and gives us the annoyingly
confusing issue  where the namespace from the upgrade workflows conflict
with the setup workflows described by this issue:

https://github.com/camunda/camunda-platform-helm/issues/3558

This PR will drop any extra metadata that 3675 introduced into the
namespace name.