refactor: deduplicate local cluster - KIND between scenario types
I was finding it difficult to work with the local cluster - KIND
workflows because there's a lot of jobs that were marked as Cancelled
and I didn't know why. The reason why is that they are duplicates, and
the original job is running on a different line.

Therefore, I separated the KIND workflows from the matrix that splits up
the different test scenarios