fix: set zeebe freespace config to application defaults
For 8.8 processing and replication limits were accidentally flipped with 79740cf .

For 8.7 previously limits were applied by env vars, which were ineffective if users set their own env vars in the values.yaml.
Thus in such cases the application defaults applied.
With 79740cf the 8.7 chart was changed to move the free space config into the configmap.
In all cases where the env vars were not effective, this was a breaking change as minimum freespace increased from from the app default of 2GB to 3GB for processing and from 1GB app default to 2GB. for replication.